requirements.txt
pandas
numpy
beautifulsoup4
requests
requests-html
lxml
lxml_html_clean
selenium
openai
python-dotenv
elasticsearch
sentence-transformers